+++ hutch_python/tests/test_utils.py
@@ -1,7 +1,17 @@
+import asyncio
 import logging
+import os
+import threading
+import time as ttime
+from signal import SIGINT, SIGQUIT
 
+import bluesky.plan_stubs as bps
 import IPython.lib.pretty as pretty
 import pytest
+from bluesky import RunEngine
+from bluesky.plan_stubs import abs_set
+from bluesky.preprocessors import finalize_wrapper
+from bluesky.utils import RunEngineInterrupted
 
 from hutch_python import utils
 
@@ -119,3 +129,81 @@ def test_hutch_banner():
     logger.debug('test_hutch_banner')
     utils.hutch_banner()
     utils.hutch_banner('mfx')
+
+
+@pytest.fixture(scope='function')
+def RE_abort():
+    """ RunEngine with extra context_managers """
+    loop = asyncio.new_event_loop()
+    loop.set_debug(True)
+    RE = RunEngine({}, loop=loop,
+                   context_managers=[utils.AbortSigintHandler,
+                                     utils.SigquitHandler])
+
+    yield RE
+
+    if RE.state != 'idle':
+        RE.halt()
+
+
+@skip_if_win32_generic
+def test_sigint_RE(RE_abort):
+    # get pid so we can send SIGINT to it specifically
+    pid = os.getpid()
+
+    def wait_plan():
+        # arbitrarily long wait time
+        yield from bps.sleep(10)
+
+    def sim_kill():
+        ttime.sleep(0.05)
+        os.kill(pid, SIGINT)
+
+    # send sigint in a different thread
+    timer = threading.Timer(0.5, sim_kill)
+    timer.start()
+
+    with pytest.raises(RunEngineInterrupted):
+        RE_abort(wait_plan())
+
+    assert RE_abort.state == 'idle'
+    assert RE_abort._exit_status == 'success'
+
+
+@skip_if_win32_generic
+def test_sigquit_two_hits(RE_abort):
+    import time
+
+    from ophyd.sim import motor
+    motor.delay = .5
+
+    pid = os.getpid()
+
+    def sim_kill(n):
+        for j in range(n):
+            time.sleep(.05)
+            os.kill(pid, SIGQUIT)
+
+    lp = RE_abort.loop
+    motor.loop = lp
+
+    def self_sig_int_plan():
+        # three hits will quit pytest
+        threading.Timer(.05, sim_kill, (2,)).start()
+        yield from abs_set(motor, 1, wait=True)
+
+    start_time = ttime.time()
+    with pytest.raises(RunEngineInterrupted):
+        RE_abort(finalize_wrapper(self_sig_int_plan(),
+                                  abs_set(motor, 0, wait=True)))
+    end_time = ttime.time()
+
+    assert RE_abort.state == 'paused'
+    # not enough time for motor to cleanup, but long enough to start
+    assert 0.05 < end_time - start_time < 0.3
+    RE_abort.abort()  # now cleanup
+
+    done_cleanup_time = ttime.time()
+    # this should be 0.5 (the motor.delay) above, leave sloppy for CI
+    assert 0.3 < done_cleanup_time - end_time < 0.6
+    assert RE_abort.state == 'idle'