+++ hutch_python/utils.py
@@ -7,8 +7,10 @@ import functools
 import inspect
 import logging
 import os
+import signal
 import socket
 import sys
+import threading
 import time
 from contextlib import contextmanager
 from functools import partial
@@ -18,6 +20,7 @@ from types import SimpleNamespace
 
 import prettytable
 import pyfiglet
+from bluesky.utils import SignalHandler
 
 from .constants import (CLASS_SEARCH_PATH, CUR_EXP_SCRIPT, HUTCH_COLORS,
                         SUCCESS_LEVEL)
@@ -433,3 +436,77 @@ def get_fully_qualified_domain_name():
         )
         logger.debug("getfqdn failed", exc_info=True)
         return ""
+
+
+abort_msg = """
+Your RunEngine has been interrupted.  If you aborted the current run (Ctrl+C),
+the RunEngine is ready for use.  If you paused the run (Ctrl+\\), you must
+resume, abort, stop, halt the RunEngine before continuing.
+"""
+
+
+class AbortSigintHandler(SignalHandler):
+    """
+    Interprets SIGINT (Ctrl+C) as an abort request.
+    """
+    def __init__(self, RE):
+        super().__init__(signal.SIGINT, log=RE.log)
+        self.RE = RE
+
+    def handle_signals(self):
+        # Check for pause requests from keyboard.
+        if self.RE.state.is_running:
+            self.RE.stop()
+            # Use our own logging
+            logger.warning("Aborting current run.")
+
+
+class SigquitHandler(SignalHandler):
+    """
+    Interprets SIGQUIT (Ctrl+\\) as standard RunEngine interrupt behavior.
+    Copied verbatim from bluesky.utils.SigintHandler, lacking a way to
+    subclassing the original while changing the signal type.
+    """
+    def __init__(self, RE):
+        super().__init__(signal.SIGQUIT, log=RE.log)
+        self.RE = RE
+        self.last_sigint_time = None  # time most recent SIGQUIT was processed
+        self.count = 0  # count SIGQUITs processed
+
+    def __enter__(self):
+        return super().__enter__()
+
+    def handle_signals(self):
+        # Check for pause requests from keyboard.
+        # TODO, there is a possible race condition between the two
+        # pauses here
+        if not self.RE.state.is_running or self.RE._interrupted:
+            return
+
+        if (self.last_sigint_time is None or
+                time.monotonic() - self.last_sigint_time > 10):
+            # reset the counter to 1
+            # It's been 10 seconds since the last SIGQUIT. Reset.
+            self.count = 1
+            if self.last_sigint_time is not None:
+                logger.debug("It has been 10 seconds since the "
+                             "last SIGQUIT. Resetting SIGQUIT "
+                             "handler.")
+            # weeee push these to threads to not block the main thread
+            threading.Thread(target=self.RE.request_pause,
+                             args=(True,)).start()
+            logger.warning("A 'deferred pause' has been requested. "
+                           "The RunEngine will pause at the next "
+                           "checkpoint. To pause immediately, hit "
+                           "Ctrl+\\ again in the next 10 seconds.")
+
+            self.last_sigint_time = time.monotonic()
+        elif self.count == 2:
+            logger.warning('trying a second time')
+            # - Ctrl+\ twice within 10 seconds -> hard pause
+            logger.debug("RunEngine detected two SIGQUITs. "
+                         "A hard pause will be requested.")
+
+            threading.Thread(target=self.RE.request_pause,
+                             args=(False,)).start()
+        self.last_sigint_time = time.monotonic()